+++ packages/webapp/src/single/MediaViewPage.ts
@@ -12,5 +12,9 @@ export type MediaViewDisableFlags = (("detail" | "map" | "similar" | "annotation
  */
 export interface MediaViewPage {
   disabled?: MediaViewDisableFlags;
+  /**
+   * Start the video playback when a video is opened. Browsers can still block the autoplay until the site has user interaction
+   */
+  autoPlayVideo?: boolean;
   [k: string]: unknown;
 }